fix(mono): extract common test helpers and fix bugs
